Update Player.dart
Fixed a typo

+++ lib/classes/Player.dart
@@ -36,7 +36,7 @@ class Player {
     return name;
   }
 
-  double get getSkill {
+  int get getSkill {
     return skill;
   }
 }
